add a bomb_bg to reimuA
