+++ src/plrmodes/reimu.h
@@ -20,3 +20,5 @@ double reimu_common_property(Player *plr, PlrProperty prop);
 void reimu_common_shot(Player *plr, int dmg);
 int reimu_common_ofuda(Projectile *p, int t);
 void reimu_common_draw_yinyang(Enemy *e, int t, const Color *c);
+void reimu_common_bomb_bg(Player *p, float alpha);
+void reimu_common_bomb_buffer_init(void);
